added logic for updating user account balance

+++ src/main/java/com/pahappa/systems/pettycashsystem/spring/dao/UserDAO.java
@@ -27,7 +27,7 @@ public class UserDAO {
     }
 
     // Read operation: Get user by ID
-    public User getUserById(int userId) {
+    public User getUserById(Long userId) {
         return getCurrentSession().get(User.class, userId);
     }
